[move][#34] 애니메이션 함수 extension으로 분리
